+++ qml/qml.qrc
@@ -1,32 +1,32 @@
 <RCC>
-    <qresource lang="en" prefix="/translations">
+    <qresource prefix="/translations" lang="en">
         <file alias="QOpenHD.qm">../translations/QOpenHD_en.qm</file>
         <file>ui/configpopup/ConfigPopup.qml</file>
         <file>qml.qrc</file>
         <file>ui/widgets/MessageHUDForm.ui.qml</file>
     </qresource>
-    <qresource lang="de" prefix="/translations">
+    <qresource prefix="/translations" lang="de">
         <file alias="QOpenHD.qm">../translations/QOpenHD_de.qm</file>
     </qresource>
-    <qresource lang="ru" prefix="/translations">
+    <qresource prefix="/translations" lang="ru">
         <file alias="QOpenHD.qm">../translations/QOpenHD_ru.qm</file>
     </qresource>
-    <qresource lang="es" prefix="/translations">
+    <qresource prefix="/translations" lang="es">
         <file alias="QOpenHD.qm">../translations/QOpenHD_es.qm</file>
     </qresource>
-    <qresource lang="fr" prefix="/translations">
+    <qresource prefix="/translations" lang="fr">
         <file alias="QOpenHD.qm">../translations/QOpenHD_fr.qm</file>
     </qresource>
-    <qresource lang="nl" prefix="/translations">
+    <qresource prefix="/translations" lang="nl">
         <file alias="QOpenHD.qm">../translations/QOpenHD_nl.qm</file>
     </qresource>
-    <qresource lang="ro" prefix="/translations">
+    <qresource prefix="/translations" lang="ro">
         <file alias="QOpenHD.qm">../translations/QOpenHD_ro.qm</file>
     </qresource>
-    <qresource lang="it" prefix="/translations">
+    <qresource prefix="/translations" lang="it">
         <file alias="QOpenHD.qm">../translations/QOpenHD_it.qm</file>
     </qresource>
-    <qresource lang="zh" prefix="/translations">
+    <qresource prefix="/translations" lang="zh">
         <file alias="QOpenHD.qm">../translations/QOpenHD_zh.qm</file>
     </qresource>
     <qresource prefix="/">
@@ -218,7 +218,7 @@
         <file>resources/master2.png</file>
         <file>resources/master3.png</file>
         <file>resources/master4.png</file>
-        <file>ui/configpopup/Eepanel.qml</file>
+        <file>ui/configpopup/EasterEggPanel.qml</file>
         <file>ui/widgets/AirspeedTempWidget.qml</file>
         <file>ui/widgets/AirspeedTempWidgetForm.ui.qml</file>
         <file>ui/widgets/AoaWidget.qml</file>
